Fix narrowing-initializer warnings with gcc 4.7.  No bug, r=themaid
--HG--
extra : rebase_source : a5cf16fbc9eb408a9eb286a2f0bb6194835a4bc2

+++ js/src/jsarray.cpp
@@ -2279,7 +2279,7 @@ js::array_sort(JSContext *cx, unsigned argc, Value *vp)
                 if (!strElements.reserve(2 * n))
                     return false;
 
-                int cursor = 0;
+                size_t cursor = 0;
                 for (size_t i = 0; i < n; i++) {
                     if (!JS_CHECK_OPERATION_LIMIT(cx))
                         return false;
